refactor: 경험 보드 선택 상태 로직 분리
